Remove CoordinatorData
