+++ src/coordination/CMakeLists.txt
@@ -9,13 +9,13 @@ target_sources(mg-coordination
         include/coordination/coordinator_config.hpp
         include/coordination/coordinator_exceptions.hpp
         include/coordination/coordinator_slk.hpp
-        include/coordination/coordinator_data.hpp
-        include/coordination/constants.hpp
+        include/coordination/coordinator_instance.hpp
         include/coordination/coordinator_cluster_config.hpp
         include/coordination/coordinator_handlers.hpp
-        include/coordination/coordinator_instance.hpp
+        include/coordination/constants.hpp
         include/coordination/instance_status.hpp
         include/coordination/replication_instance.hpp
+        include/coordination/raft_instance.hpp
 
         include/nuraft/coordinator_log_store.hpp
         include/nuraft/coordinator_state_machine.hpp
@@ -26,10 +26,10 @@ target_sources(mg-coordination
         coordinator_state.cpp
         coordinator_rpc.cpp
         coordinator_server.cpp
-        coordinator_data.cpp
-        coordinator_instance.cpp
         coordinator_handlers.cpp
+        coordinator_instance.cpp
         replication_instance.cpp
+        raft_instance.cpp
 
         coordinator_log_store.cpp
         coordinator_state_machine.cpp
