+++ src/query/interpreter.hpp
@@ -105,13 +105,13 @@ class CoordinatorQueryHandler {
   };
 
   /// @throw QueryRuntimeException if an error ocurred.
-  virtual void RegisterInstance(const std::string &coordinator_socket_address,
-                                const std::string &replication_socket_address,
-                                const std::chrono::seconds instance_check_frequency, const std::string &instance_name,
-                                CoordinatorQuery::SyncMode sync_mode) = 0;
+  virtual void RegisterReplicationInstance(const std::string &coordinator_socket_address,
+                                           const std::string &replication_socket_address,
+                                           const std::chrono::seconds instance_check_frequency,
+                                           const std::string &instance_name, CoordinatorQuery::SyncMode sync_mode) = 0;
 
   /// @throw QueryRuntimeException if an error ocurred.
-  virtual void SetInstanceToMain(const std::string &instance_name) = 0;
+  virtual void SetReplicationInstanceToMain(const std::string &instance_name) = 0;
 
   /// @throw QueryRuntimeException if an error ocurred.
   virtual std::vector<coordination::InstanceStatus> ShowInstances() const = 0;
